Fix clear()/empty() confusion and doubles being printed as ints.
Also add a note in src/stats/coordinate_gen about random_shuffle
not necessarily using uint64_t.

And comment out other variables for excluded g++ on windows target in CI
workflow.

+++ src/singlewinner/pairwise/simple_methods.cc
@@ -101,7 +101,7 @@ std::pair<ordering, bool> ext_minmax::pair_elect(const abstract_condmat &
 		if (debug) {
 			std::cout << "Before sorting: " << counter << "\t";
 			copy(beat_data.first.begin(), beat_data.first.end(),
-				std::ostream_iterator<int>(std::cout, "   "));
+				std::ostream_iterator<double>(std::cout, "   "));
 			std::cout << std::endl;
 		}
